% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pandoc_to}
\alias{pandoc_to}
\alias{pandoc_from}
\title{Get current Pandoc input and output formats for a Rmarkdown document}
\usage{
pandoc_to(to)

pandoc_from()
}
\arguments{
\item{to}{A character vector of output formats to be checked against the
current Pandoc output format. If provided, \code{pandoc_to()} will check
the current output format against the provided \code{to}.}
}
\value{
The current input or output Pandoc format used. Or \code{TRUE} for
  \code{pandoc_to(to)} if the current output format is one of \code{to}
}
\description{
These functions are to be used with R Markdown documents. \code{pandoc_to()} returns
the Pandoc output format and \code{pandoc_from()} returns Pandoc input
format. \code{pandoc_to(to)} allows to check the current output format
against a set of format names.
}
\details{
In addition to \code{knitr::\link{is_latex_output}} and
\code{knitr::\link{is_htmlt_output}}, these functions may be useful for
conditional output of more specific output formats.

Internally, the Pandoc output format of the current R Markdown document is
stored in \code{knitr::\link{opts_knit}$get('rmarkdown.pandoc.to')}, and the
Pandoc input format in
\code{knitr::\link{opts_knit}$get('rmarkdown.pandoc.from')}
}
\note{
See available Pandoc formats, in
  \href{https://pandoc.org/MANUAL.html}{Pandoc's Manual}
}
\examples{
# Get current formats
knitr::pandoc_from()
knitr::pandoc_to()
# Test if current output format is 'docx'
knitr::pandoc_to("docx")
}
\seealso{
\code{knitr::\link{is_latex_output}},
  \code{knitr::\link{is_htmlt_output}}
}
