% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_input}
\alias{current_input}
\title{Query the current input filename}
\usage{
current_input(dir = FALSE)
}
\arguments{
\item{dir}{Whether to prepend the current working directory to the file
path, i.e., whether to return an absolute path or a relative path.}
}
\value{
A character string, if this function is called inside an input
document. Otherwise \code{NULL}.
}
\description{
Returns the name of the input file passed to \code{\link[=knit]{knit()}}.
}
