% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{eng_r}
\alias{eng_r}
\title{Engine for R}
\usage{
eng_r(options)
}
\arguments{
\item{options}{A list of chunk options. Usually this is just the object
\code{options} associated with the current code chunk.}
}
\description{
This function handles the execution of R code blocks (when the chunk option \code{engine} is  \code{'R'})
and generates the R output for each code block.
}
\details{
This engine function has one argument \code{options}: the source code of the current chunk  is in \code{options$code}.
It returns a processed output which can consist of data frames (as tables), graphs or character output.
This function is intended for advanced use to allow developers to extend R,
and customise the pipeline with which R code is executed and processed within knitr.
}
