% Please edit documentation in R/engine-V8.R
\name{engine_V8}
\alias{engine_V8}
\title{Javascript V8 engine (require the V8 package)}
\usage{
engine_V8(options)
}
\arguments{
\item{options}{a list of chunk options}
}
\description{
example at \link{https://gist.github.com/stla/10718cda54bc5c0aa6fb}
}
