% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyBashMarkdown.R
\name{formatBash}
\alias{formatBash}
\title{This is used to format bash chunk code in rmarkdown.}
\usage{
formatBash(code, width.cutoff = getOption("width"), ...)
}
\arguments{
\item{code:}{input code}
}
\value{
Output a vector of codes. Each element is a line in the md file.
}
\description{
This is used to format bash chunk code in rmarkdown.
}
\examples{

}
