% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks-html.R, R/hooks-asciidoc.R,
%   R/hooks-latex.R, R/hooks-md.R, R/hooks-rst.R, R/hooks-textile.R,
%   R/hooks-typst.R
\name{hook_plot_html}
\alias{hook_plot_html}
\alias{hook_plot_asciidoc}
\alias{hook_plot_tex}
\alias{hook_plot_md}
\alias{hook_plot_rst}
\alias{hook_plot_textile}
\alias{hook_plot_typst}
\title{Default plot hooks for different output formats}
\usage{
hook_plot_html(x, options)

hook_plot_asciidoc(x, options)

hook_plot_tex(x, options)

hook_plot_md(x, options)

hook_plot_rst(x, options)

hook_plot_textile(x, options)

hook_plot_typst(x, options)
}
\arguments{
\item{x}{Filename for the plot (a character string).}

\item{options}{A list of the current chunk options.}
}
\value{
A character string of code, with plot filenames wrapped.
}
\description{
These hook functions define how to mark up graphics output in different
output formats.
}
\details{
Depending on the options passed over, \code{hook_plot_tex} may return the
normal \samp{\\includegraphics{}} command, or \samp{\\input{}} (for tikz
files), or \samp{\\animategraphics{}} (for animations); it also takes many
other options into consideration to align plots and set figure sizes, etc.
Similarly, \code{hook_plot_html}, \code{hook_plot_md} and
\code{hook_plot_rst} return character strings which are HTML, Markdown, reST
code.

For animations (i.e. when the chunk option \code{fig.show} is \code{'animate'}),
\code{hook_plot_tex} generates \samp{\\animategraphics{}} unless a hook function
has been provided via the chunk option \code{animation.hook} or the package option
\code{animation.fun}, in which case that function generates the LaTeX code instead.
It is called only once per chunk, with the filename of the last plot. The
built-in hooks (e.g. \code{\link{hook_ffmpeg_html}}) generate HTML, and are
ignored for LaTeX output.

In most cases we do not need to call these hooks explicitly, and they were
designed to be used internally. Sometimes we may not be able to record R
plots using \code{grDevices::\link{recordPlot}()}, and we can make use of
these hooks to insert graphics output in the output document; see
\code{\link{hook_plot_custom}} for details.
}
\examples{
# this is what happens for a chunk like this

# <<foo-bar-plot, dev='pdf', fig.align='right'>>=
hook_plot_tex("foo-bar-plot.pdf", opts_chunk$merge(list(fig.align = "right")))

# <<bar, dev='tikz'>>=
hook_plot_tex("bar.tikz", opts_chunk$merge(list(dev = "tikz")))

# <<foo, dev='pdf', fig.show='animate', interval=.1>>=

# 5 plots are generated in this chunk
hook_plot_tex("foo5.pdf", opts_chunk$merge(list(fig.show = "animate", interval = 0.1,
    fig.cur = 5, fig.num = 5)))
}
\references{
\url{https://yihui.org/knitr/hooks/}
}
\seealso{
\code{\link{hook_plot_custom}}
}
