% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{html_to_r}
\alias{html_to_r}
\title{Extract R code from a knitted R Markdown HTML file.}
\usage{
html_to_r(rmarkdown_html, padding, inc_out)
}
\arguments{
\item{rmarkdown_html}{The  HTML output  of a knitted Rmd file}

\item{padding}{Specifies what goes between the last character of a code block and the
first character of the next code block. Defaults to two newlines (which gives
the visual appearance of one newline between code blocks).}

\item{inc_out}{\code{TRUE}/\code{FALSE} as to whether to include output of code chucks. Defaults
to \code{TRUE}.}
}
\value{
A character vector of length 1 containing the R code extracted from the
    R Markdown HTML file.
}
\description{
Extract R code from a knitted R Markdown HTML file.
}
\examples{
cat((rmarkdown_html <- "<html><body>Intro
```R\n1 * 1\n```\nmore text\n```r\n2 * 2\n```
some more text\n```\n3 * 3\n```\nThe End.
</body></html>\n"))

html_to_r(rmarkdown_html)

cat(html_to_r(rmarkdown_html))

}
\author{
Steve Condylios
}
