% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{imgur_upload}
\alias{imgur_upload}
\title{Upload an image to imgur.com}
\usage{
imgur_upload(file, key = xfun::env_option("knitr.imgur.key"), ...)
}
\arguments{
\item{file, key, ...}{See \code{\link[xfun:upload_imgur]{xfun::upload_imgur()}}.}
}
\description{
This function is an alias to \code{xfun::upload_imgur()}. It is kept in
\pkg{knitr} only for backward-compatibility reasons. You are recommended to
use \code{xfun::upload_imgur()} directly instead.
}
\keyword{internal}
