% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{knit_cnd_format}
\alias{knit_cnd_format}
\title{Format a condition prior to sewing}
\usage{
knit_cnd_format(cnd)
}
\description{
This generic is called by condition methods for
\code{\link{sew}()}. Methods should return a full condition
message, including the prefix (e.g. `Error:` or `Warning:`,
including a call if the condition includes one).
}
\keyword{internal}
