% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knit_global}
\alias{knit_global}
\title{The global environment for evaluating code}
\usage{
knit_global(envir = NULL)
}
\arguments{
\item{envir}{If \code{NULL}, the function returns the \code{envir} argument
of \code{\link[=knit]{knit()}}, otherwise it should be a new environment for
evaluating code, in which case the function returns the old environment
after setting the new environment.}
}
\description{
Get or set the environment in which code chunks are evaluated.
}
