% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{knit_params_yaml}
\alias{knit_params_yaml}
\title{Extract knit parameters from YAML text}
\usage{
knit_params_yaml(yaml, evaluate = TRUE)
}
\arguments{
\item{yaml}{Character vector containing the YAML text.}

\item{evaluate}{If \code{TRUE} (the default), expression values embedded
within the YAML will be evaluated. If \code{FALSE}, parameters defined with
an expression will have the parsed but unevaluated expression in their
\code{value} field.}
}
\value{
List of objects of class \code{knit_param} that correspond to the
parameters declared in the \code{params} section of the YAML. See
\code{\link[=knit_params]{knit_params()}} for a full description of these objects.
}
\description{
This function reads the YAML front-matter that has already been extracted
from a document and returns a list of any parameters declared there.
}
\seealso{
\code{\link[=knit_params]{knit_params()}}
}
