% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_latex}
\alias{raw_latex}
\title{Mark character strings as LaTeX}
\usage{
raw_latex(x, ...)
}
\arguments{
\item{x}{The character vector to be protected.}

\item{...}{Arguments to be passed to \code{\link{asis_output}()}.}
}
\description{
If possible, this function wraps content in a LaTeX raw attribute block
If possible, this function wraps content in a LaTeX raw attribute block
which protects it from being escaped by pandoc. See
(see \url{https://pandoc.org/MANUAL.html#generic-raw-attribute}).
}
\examples{
raw_latex('\\\\emph{some text}')
}
\seealso{
raw_output
}
