% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversion.R
\name{rst2pdf}
\alias{rst2pdf}
\title{A wrapper for rst2pdf}
\usage{
rst2pdf(input, command = "rst2pdf", options = "")
}
\arguments{
\item{input}{The input rst file.}

\item{command}{Character string giving the path of the
\command{rst2pdf} program. If the program is not in your PATH, the full path has to be
given here.}

\item{options}{Extra command line options, e.g., \code{'-v'}.}
}
\value{
An input file \file{*.rst} will produce \file{*.pdf} and this output
filename is returned if the conversion was successful.
}
\description{
Convert reST to PDF using \command{rst2pdf} (which converts from rst to PDF
using the ReportLab open-source library).
}
\references{
\url{https://github.com/rst2pdf/rst2pdf}
}
\seealso{
\code{\link[=knit2pdf]{knit2pdf()}}
}
\author{
Alex Zvoleff and Yihui Xie
}
