% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{sew}
\alias{sew}
\title{Wrap evaluated results for output}
\usage{
sew(x, options = list(), ...)
}
\arguments{
\item{x}{Output from \code{\link[evaluate:evaluate]{evaluate::evaluate()}}.}

\item{options}{A list of chunk options used to control output.}

\item{...}{Other arguments to pass to methods.}
}
\description{
This function is mainly for internal use: it is called on each part of the
output of the code chunk (code, messages, text output, and plots, etc.) after
all statements in the code chunk have been evaluated, and will sew these
pieces of output together into a character vector.
}
