% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{stitch}
\alias{stitch}
\alias{stitch_rhtml}
\alias{stitch_rmd}
\title{Automatically create a report based on an R script and a template}
\usage{
stitch(
  script,
  template = system.file("misc", "knitr-template.Rnw", package = "knitr"),
  output = NULL,
  text = NULL,
  envir = parent.frame()
)

stitch_rhtml(..., envir = parent.frame())

stitch_rmd(..., envir = parent.frame())
}
\arguments{
\item{script}{Path to the R script.}

\item{template}{Path of the template to use. By default, the Rnw template in
this package; there is also an HTML template in \pkg{knitr}.}

\item{output}{Output filename, passed to \code{\link[=knit]{knit()}}). By default,
the base filename of the script is used.}

\item{text}{A character vector. This is an alternative way to provide the
input file.}

\item{envir}{Environment in which code chunks are to be evaluated, for
example, \code{\link[=parent.frame]{parent.frame()}}, \code{\link[=new.env]{new.env()}}, or
\code{\link[=globalenv]{globalenv()}}).}

\item{...}{Arguments passed to \code{stitch()}.}
}
\value{
path of the output document
}
\description{
This is a convenience function for small-scale automatic reporting based on
an R script and a template. The default template is an Rnw file (LaTeX);
\code{stitch_rhtml()} and \code{stitch_rmd()} are wrappers on top of
\code{stitch()} using the R HTML and R Markdown templates respectively.
}
\details{
The first two lines of the R script can contain the title and author of the
report in comments of the form \verb{## title:} and \verb{## author:}. The
template must have a token \verb{\\\%sCHUNK_LABEL_HERE}, which will be used to
input all the R code from the script. See the examples below.

The R script may contain chunk headers of the form
\verb{## ---- label, opt1=val1, opt2=val2}, which will be copied to the template; if no chunk
headers are found, the whole R script will be inserted into the template as
one code chunk.
}
\examples{
s = system.file("misc", "stitch-test.R", package = "knitr")
if (interactive()) stitch(s)  # compile to PDF

# HTML report
stitch(s, system.file("misc", "knitr-template.Rhtml", package = "knitr"))

# or convert markdown to HTML
stitch(s, system.file("misc", "knitr-template.Rmd", package = "knitr"))

unlink(c("stitch-test.html", "stitch-test.md", "figure"), recursive = TRUE)
}
\seealso{
\code{\link[=spin]{spin()}} (turn a specially formatted R script to a report)
}
