% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-stringr_suggestions.R
\name{use_stringr}
\alias{use_stringr}
\title{Use stringr functions internally}
\usage{
use_stringr()
}
\value{
\code{TRUE} or \code{FALSE}: whether or not the original \code{stringr}
functions will be used.
}
\description{
Originally written using \code{stringr} functions throughout,
\code{knitr} is now designed to work without \code{stringr}. Nonetheless,
\code{stringr} is used by default. To force \code{knitr} to not use \code{stringr},
you can set an environment variable:

\preformatted{Sys.setenv("KNITR_USE_STRINGR" = "FALSE")}

or use the package option

\preformatted{options("knitr.use.stringr" = FALSE)}

The environment variable has precedence over the package option
if and only if it is set to \code{FALSE}. Otherwise, the package
option is used, with a default value of \code{TRUE}.

Our intention is to eventually switch the default, so \code{knitr} can be used
without installing \code{stringr}. In the meantime, we provide this options
so authors can test their scripts with and without stringr before we make
a potentially breaking change.

The function \code{use_stringr} returns the current use of \code{stringr}.
}
