% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_opts}
\alias{with_opts}
\title{Reference label with chunk options}
\usage{
with_opts(ref.label)
}
\arguments{
\item{ref.label}{A string to be passed to the `ref.label` chunk option.}
}
\description{
When the \code{ref.label} chunk option is specified with this function,
the chunk inherits chunk options from the reference chunk. Note that
explicitly specified chunk options have priority over the inherited chunk
options.
}
