% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{wrap}
\alias{wrap}
\title{Wrap evaluated results for output}
\usage{
wrap(x, options = list(), ...)
}
\arguments{
\item{x}{Output from \code{evaluate::\link{evaluate}()}.}

\item{options}{A list of chunk options used to control output.}

\item{...}{Other arguments to pass to methods.}
}
\description{
This function is mainly for internal use: it is called on each part of the
output of the code chunk (code, messages, text output, and plots, etc.) after
all statements in the code chunk have been evaluated.
}
