% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{write_bib}
\alias{write_bib}
\title{Generate BibTeX bibliography databases for R packages}
\usage{
write_bib(..., prefix = getOption("knitr.bib.prefix", "R-"))
}
\arguments{
\item{..., prefix}{Arguments passed to \code{\link[xfun:pkg_bib]{xfun::pkg_bib()}}.}
}
\description{
A wrapper function of \code{xfun::pkg_bib()}.
}
