\docType{data}
\name{all_patterns}
\alias{all_patterns}
\title{All built-in patterns}
\format{List of 7
 $ rnw     :List of 7
 $ brew    :List of 1
 $ tex     :List of 8
 $ html    :List of 5
 $ md      :List of 4
 $ rst     :List of 5
 $ asciidoc:List of 6}
\usage{
all_patterns
}
\description{
  This object is a named list of all built-in patterns.
}
\examples{
all_patterns$rnw
all_patterns$html

str(all_patterns)
}
\references{
  Usage: \url{http://yihui.name/knitr/patterns}
}
\keyword{datasets}
