% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\docType{data}
\name{all_patterns}
\alias{all_patterns}
\title{All built-in patterns}
\format{An object of class \code{list} of length 8.}
\usage{
all_patterns
}
\description{
This object is a named list of all built-in patterns.
}
\examples{
all_patterns$rnw
all_patterns$html

str(all_patterns)
}
\references{
Usage: \url{https://yihui.org/knitr/patterns/}
}
\seealso{
\code{\link{knit_patterns}}
}
\keyword{datasets}
