% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_words}
\alias{combine_words}
\title{Combine multiple words into a single string}
\usage{
combine_words(...)
}
\arguments{
\item{...}{Arguments passed to \code{xfun::\link[xfun]{join_words}()}.}
}
\description{
This is a wrapper function of \code{xfun::join_words()}.
}
