% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{is_low_change}
\alias{is_low_change}
\title{Compare two recorded plots}
\usage{
is_low_change(p1, p2)
}
\arguments{
\item{p1, p2}{Plot results}
}
\value{
Logical value indicating whether `p2` is a low-level update of `p1`.
}
\description{
Compare two recorded plots
}
