% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knit_global}
\alias{knit_global}
\title{The global environment in which code chunks are evaluated}
\usage{
knit_global()
}
\description{
This function makes the environment of a code chunk accessible inside a
chunk.
}
\details{
It returns the \code{envir} argument of \code{\link{knit}}, e.g. if we call
\code{\link{knit}()} in the global environment, \code{knit_global()} returns
R's global environment by default. You can call functions like
\code{\link{ls}()} on this environment.
}
