% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{knit_hooks}
\alias{knit_hooks}
\title{Hooks for R code chunks, inline R code and output}
\format{
An object of class \code{list} of length 7.
}
\usage{
knit_hooks
}
\description{
A hook is a function of a pre-defined form (arguments) that takes values of
arguments and returns desired output. The object \code{knit_hooks} is used to
access or set hooks in this package.
}
\examples{
knit_hooks$get('source'); knit_hooks$get('inline')
}
\references{
Usage: \url{https://yihui.org/knitr/objects/}

Components in \code{knit_hooks}: \url{https://yihui.org/knitr/hooks/}
}
\keyword{datasets}
