\docType{package}
\name{knitr-package}
\alias{knitr}
\alias{knitr-package}
\title{A general-purpose tool for dynamic report generation in R}
\description{
  This is an alternative tool to Sweave with more flexible
  design and new features like cache and fine control of
  graphics. It is not limited to LaTeX and is ready to be
  customized to process other file formats. See the package
  website in the references for more information and
  examples.
}
\author{
  Yihui Xie <\url{http://yihui.name}>
}
\references{
  Full documentation and demos:
  \url{http://yihui.github.com/knitr/}; FAQ's:
  \url{https://github.com/yihui/knitr/blob/master/FAQ.md}
}
