\docType{package}
\name{knitr-package}
\alias{knitr}
\alias{knitr-package}
\title{A general-purpose tool for dynamic report generation in R}
\description{
  This is an alternative tool to Sweave with more flexible
  design and new features like cache and fine control of
  graphics. It is not limited to LaTeX and is ready to be
  customized to process other file formats. See the package
  website in the references for more information and
  examples.
}
\note{
  The pronunciation of \pkg{knitr} is similar to
  \emph{neater} (neater than what?) or you can think of
  \emph{knitter} (but it is \emph{single t}).  The name
  comes from \code{knit} + \code{R} (while \code{Sweave} =
  \code{S} + \code{weave}).
}
\author{
  Yihui Xie <\url{http://yihui.name}>
}
\references{
  Full documentation and demos:
  \url{http://yihui.name/knitr/}; FAQ's:
  \url{https://github.com/yihui/knitr/blob/master/FAQ.md}
}
\seealso{
  The core function in this package: \code{\link{knit}}. If
  you are an Sweave user, see \code{\link{Sweave2knitr}} on
  how to convert Sweave files to \pkg{knitr}.
}
