% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{partition_chunk}
\alias{partition_chunk}
\title{Partition chunk options from the code chunk body}
\usage{
partition_chunk(engine, code)
}
\description{
This is a wrapper function calling \code{xfun::\link[xfun]{divide_chunk}()}
under the hood.
}
\keyword{internal}
