% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{partition_chunk}
\alias{partition_chunk}
\title{Partition YAML options from a code chunk}
\usage{
partition_chunk(engine, code)
}
\arguments{
\item{engine}{the lower-case name of the knitr engine to run (to determine 
the appropriate comment character).}

\item{code}{a character vector where each element represents one line of code.}
}
\value{
a list with the following items:
  \describe{
    \item{\code{options}}{The parsed options (if any) as a list}
    \item{\code{src}}{The part of the input that contains the options}
    \item{\code{code}}{The part of the input that contains the code}
  }
}
\description{
Partition YAML options from a code chunk
}
\examples{
# parse yaml-like items
yaml_like <- c("#| label: mine", "#| echo: true", "#| fig.width: 8", "#| foo: bar", "1 + 1")
writeLines(yaml_like)
partition_chunk("r", yaml_like)

# parse CSV syntax
csv_like <- c("#| mine, echo = TRUE, fig.width = 8, foo = 'bar'", "1 + 1")
writeLines(csv_like)
partition_chunk("r", csv_like)
}
\keyword{internal}
