% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_crop}
\alias{plot_crop}
\title{Crop a plot (remove the edges) using PDFCrop or ImageMagick}
\usage{
plot_crop(x, quiet = TRUE)
}
\arguments{
\item{x}{Filename of the plot.}

\item{quiet}{Whether to suppress standard output from the command.}
}
\value{
The original filename.
}
\description{
The command \command{pdfcrop} (often shipped with a LaTeX distribution) is
executed on a PDF plot file, and \code{magick::\link[magick]{image_trim}()}
is executed for other types of plot files.
}
\references{
PDFCrop: \url{https://www.ctan.org/pkg/pdfcrop}. If you use
  TinyTeX, you may install \command{pdfcrop} with
  \code{tinytex::tlmgr_install('pdfcrop')}.
}
