% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversion.R
\name{rnw2pdf}
\alias{rnw2pdf}
\title{Convert Rnw to PDF}
\usage{
rnw2pdf(
  input,
  output = with_ext(input, "pdf"),
  compiler = "xelatex",
  envir = parent.frame(),
  quiet = FALSE,
  clean = TRUE,
  error = FALSE,
  ...
)
}
\arguments{
\item{input}{Path to the input file.}

\item{output}{Path of the output file. By default, it uses the same name as the
\code{input}, but changes the file extension to ".pdf".}

\item{compiler}{The LaTeX engine that be passed to \code{tinytex::\link{latexmk}()}.
The default engine is \code{xelatex}.}

\item{envir}{Environment in which code chunks are to be evaluated, for
example, \code{\link{parent.frame}()}, \code{\link{new.env}()}, or
\code{\link{globalenv}()}).}

\item{quiet}{Boolean; suppress the progress bar and messages?}

\item{clean}{When \code{TRUE}, the intermediate files would be removed}

\item{error}{When \code{FALSE}, it stops when any error happens in knitting}

\item{...}{Options to be passed to \code{tinytex::\link[tinytex]{latexmk}()}}
}
\description{
This is a wrapper of \code{\link{knit2pdf}()}, with differences that:
}
