% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{wrap}
\alias{wrap}
\title{Wrap evaluated results for output}
\usage{
wrap(x, options = list(), ...)
}
\arguments{
\item{x}{output from \code{evaluate::\link{evaluate}()}}

\item{options}{List of options used to control output}

\item{...}{Other arguments to pass to methods}
}
\description{
This function is mainly for internal use:  it is called
on each part of the output of the code chunk
(code, messages, output, plots, etc.) after all
statements in the code chunk have been evaluated,
and allows final processing.
}
